fix: :bug: moved card to components and used next/image
